This (partially) fixes CLARIAH/WP4-Requirements#59.
We now first retrieve all predicates for triples in which a resource appears, then send out queries (in separate threads) for each of those triples, and rejoin them later. The limit is evenly distributed across all predicates (apart from the results in which the resource is itself a predicate: those will always get 1/3 of the limit).

Why partially fixed? Because we do not yet do this for CONSTRUCT queries (i.e. RDF/Turtle lookups). Not sure whether a LIMIT makes sense here anyway.